Added support for single aggregate
A single aggregate can now be used, together with granularity. Added aggregates and granularity as options, transformed key of resultant response-value to same key as for no aggregates (as that is what cdp-spark-connector supports) (this is why only one aggregate is supported) and worked around bug with getting timestamps before start time.