Fix upsert for Events
The old implementation didn't consider the fact that if there are conflicts the rest of the events will not be processed by CDP.

Quote from CDP documentation:
Responds with a 409 (CONFLICT) if the source & sourceId already exists. It will not proceed if there is a conflict, so no events will be created in that case. The JSON body contains a field called 'duplicates', which is a list of conflicting items, where each item has an id, source and sourceId.