Optimize pushdown for expressions involving and
Currently spark queries involving and are performed as two separate
requests to the API even though it could be performed as one request
with several query parameters. Complex expressions involving several
nested ands and ors are also optimized by pushing the ands to the bottom
while multiplying the requests on both sides of the and by each other.