Dangerous to have a max thread cap with SyncQueue
SyncQueue doesn't queue the Runnables in a queue collection, but instead queues the runnable for an available thread, or if no threads are available it will create a thread. However, if max number of threads is reached, it'll throw. I propose using a fixed thread pool instead based on the number of CPUs available.
I don't see why this needs to be related to max concurrent requests, as this thread pool is not used for the actual request. It's only used for scheduling a request. It could probably be smaller as well, as it shouldn't take long to schedule.