Exclude netty-transport-native-epoll from dependencies
While it would be nice, and it's perhaps something we should see if we
can support for advanced users (including ourselves), Spark by default
does not handle "native" dependencies like this, where the package is
named something like "Linux.x86" but the dependency name doesn't
include the architecture.

Spark downloads the correct dependency but then gets confused because
it had a different name from the package.