Don't use .handleError with non-exhaustive match
ApplicativeError.handleError takes a regular function (as opposed to a partial function), and expects it to always handle every error.
But we were using this function in a few places under the assumption that it accepts a partial function, so we will throw a MatchError if the exception did not match the pattern.
So let's use .adaptError when we just want to map certain exceptions to different exceptions, or .recover when we want to recover successfully from certain exceptions.