[WIP] Update to Cats Effect 3
This doesn't work, unfortunately, because Spark includes cats-kernel
2.0.0-M4, which is not binary compatible with non-milestone releases
of cats-kernel 2.x, which is required by Cats Effect 3.

It also doesn't work because circe 0.14 and later depends on a version
of shapeless that is incompatible with the one included in Spark.

This PR currently shows how to build a fat JAR to get around that, but
that's not a viable solution for production use.