Fix reading datapoints with too many partitions
Now we limit the number of ranges to 10k, otherwise we'd have to do
paging when splitting and parallelizing this large number of ranges
will probably only hurt performance.

Also, I have changed the range size to be 10 pages instead of only 1.
Paging through 10 datapoints will take ~30 seconds even when
CDF is slow, so it won't hurt performance and it will allow us to proceed
with more even ranges and with less paralelism (so hopefully less rate-limiting).

The problem is that we are only capable of handling 10k ranges
(otherwise we'd have to page through the aggregates). First, this adds
a check that when we exceed the 10k limit, we use the higher granularity.
Increasing the partition size to 10 pages should help not to hit this limit in the first place.

Other option would be to add paging to this partition listing,
however increasing parallelism beyond 10k parallel requests is pretty much
pointless when CDF allows only 100 requests per second. I understand
that we want parallelize, but paging uses the request budget more
efficiently by getting full 100k datapoint load in 9/10 of the requests.
And going through 10 pages is not going to be slow, it should be well under 30 seconds.