Fix performance problem with datapoint insertion
Main problem was that when inserting into very many time series,
we converged into making 1 request per datapoint - because
each datapoint for one series ended up in a different partition.

This lead to a surge of parallel requests -> rate limiting -> more problems.
So now we also limit the write parallism in the same way as
we do with read parallism.