Custom mapping of RAW Json -> Spark Row to get better performance
We used to use Spark's from_json UDF for this purpose,
but it turns out that it's really slow with large numbers of columns.
For this reason, I copied their code for mapping Json values,
adjusted it to use circe, so we don't need to to serialize
and parse it again.

Currently, date / time values are not supported, but AFAIK it's
impossible to use them with RAW in Jetfire.