Add cdf_dump: a CLI utility for running spark DS
Because:
* I had to debug quite a bit of Spark Data Source
issues

* I needed to get some large data sets
(at least too large for Python to hold in memory) out
of CDF for analysis

* PySpark for broken with the current DS version for some time

* PySpark is broken anytime I touch Python a bit (Conda..., sigh)

So I just wrote this tool in Scala for myself, so I can run simple SQL
Spark SQL queries ad hoc
without going through our unit tests and without worrying
if PySpark is working this time. Today, I just
added bit of documentation to the CLI options
and set up sbt assembly to produce one jar.
Not that we'd have the pressing need of maintaining another tool,
but it might be useful for internal use :)