Limit per-partition parallelism if maxWritePartitions is set
maxWritePartitions is used when ingesting into WDL to reduce the amount
of parallel requests.

When writing a dataframe to CDF, the `partitions` property is not sent
to the spark datasource.  That means that the value is set to 200 since
it is the default. The result was that each spark partition ran with a
very high `maxParallelism` setting.