Add some explicit override-s
Seeing if method is doing an override or not is quite helpful when
doing refactorings. Those that do override have to be kept as is
or refactored along with the interface, while non-overriding ones
are safe to change in any suitable way.

Sadly didn't manage to persist a linter for this
- scalafix + https://github.com/xplosunn/ForceOverrides doesn't work
  because that extension is outdated and not compatible with recent scalafix
- https://github.com/wartremover/wartremover-contrib#missingoverride
  produces warning also for `val` overrides (that's ok) but also seemingly
  some false positives at places that aren't even declarations but rather fn calls