CDF-20199: fix dms externalId.{startswith,in} filters
`externalId` is a special property and exists on `node`/`edge` rather than on
`space/versionedExternalId`.
There already was handling of that for `EqualTo` filters but was missing
for `In` and `Prefix` filters so let's extend this.

The code around here could definitely benefit from some refactoring, maybe
introducing dedicated types for various kinds of references etc.
But that needs more thought and for now mighty copy-paste it is.

[CDF-20199]