CDF-20103: make RawTableRelationTest less flaky (#877)
* CDF-20103: make RawTableRelationTest less flaky

Making the test create all the data it expects to exist so that it can be
run in any project.

Small change in behavior is `lastUpdatedTime` range test, ideally we should
be able to test that too, but we don't control the value when ingesting into
raw. So let's for now have slightly weaker test using `lastUpdatedTime(now) >= somewhere in the past`
and more strong test using range of `startTime`. Should be good enough for now,
plus there are other pushdown tests.

[CDF-20103]

* struct-test table too

* scala 2.12&2.13