Ensure a sequential write to DM by always using a single output
We want to go slow to go fast with data modeling, and never write with
more than one partition at a time. This ensures this by repartitioning
down to 1 partition before emitting the result.

Since we almost always had 200 partitions going into the write, we
would always clamp this down to the ideal, which was set to 1. But, if
we actually had less than 50 partitions going into this, it would not
be repartitioned and we would actually try to write to DMS with 50
parallel writes.

I first didn't find any way to trigger this, but I think I can trigger
it by reading directly from time series, sequences etc., which have 50
or smaller as default partitions.

This is very hard to verify in production, but there should be no
clear downside to ensuring this.