Add parameter to enable/disable shared request throttle
Retrying http client already does exponential backoff with jitter and
that should backoff and spread the load on failed requests.

BackpressureThrottlingBackend reacts to any 429/503 and throttles all
in-flight requests for a random period of time for 400..800ms.
Deficiencies:
- it is randomized but unblocks all throttled requests simultaneously
- it applies throttle to all resources even if they are unrelated

Adding an option to disable backpressure backend and rely only on
established exponential backoff algorithm.

Maybe for later we could revisit how to coordinate throttling for
requests for same resource, so when one request hits 429/503 all
new requests for that client would slow down before even trying
their luck. This needs to be randomized and improved relative
to current BackpressureThrottlingBackend, so for some other day.