Submit last (empty) sync cursor to callback
Currently, we return the cursor from the /sync DMS call as part of the
rows returned. This cursor has a max lifespan (currently 3 days), and needs
to be refreshed even if there are no rows returned.

As a workaround, for now, we therefore submit the cursor to a configured
callback URL when the /sync query returns no entries.