Add attempt number to metric names
By adding the attempt number to the metrics it will be possible to get
the highest attempt for each stage and task and allowing ignoring
metrics coming from failed/retried tasks. A good example would be the
effective number of rows added.

Though, the metrics don't allow labels, and instead uses dot-delimited
fields such as job id.
By keeping always a consistent number of fields, these can be extracted
while aggregating the metrics. Therefore, it is the responsibility of
the MetricSource to enforce this consistency.