Add TTL to metric sources
When a new metric was created (from a new jobId, for example), this
metric would be tracked forever. It would stay in memory in a concurrent
hash map, and it would stay registered with the metrics system.

With these changes, the metrics will be deregistered and removed from
the map upon one hour of inactivity. This is especially useful now that
the cardinality has increased with the tracking of stages and tasks as
part of the metrics.