Celo/cdf 21249/attempt number metrics (#924)
* Add attempt number to metric names

By adding the attempt number to the metrics it will be possible to get
the highest attempt for each stage and task and allowing ignoring
metrics coming from failed/retried tasks. A good example would be the
effective number of rows added.

Though, the metrics don't allow labels, and instead uses dot-delimited
fields such as job id.
By keeping always a consistent number of fields, these can be extracted
while aggregating the metrics. Therefore, it is the responsibility of
the MetricSource to enforce this consistency.

* Add TTL to metric sources

When a new metric was created (from a new jobId, for example), this
metric would be tracked forever. It would stay in memory in a concurrent
hash map, and it would stay registered with the metrics system.

With these changes, the metrics will be deregistered and removed from
the map upon one hour of inactivity. This is especially useful now that
the cardinality has increased with the tracking of stages and tasks as
part of the metrics.

* Revert "Add TTL to metric sources"

This reverts commit d4368f13dac525369d519cf3a435f8320d3d2817.

* Add deregistering of metrics

Upon job end, the driver may call a method to both clean the tracking
hashmap and deregister the source from the MetricsSystem.

The new method will remove all metrics belonging to a given job UUID.

* Add verbose logging to the (de)registering of metrics

This can be quite noisy, but since these metrics live internally on the
metric sink, they can provide a temporary sanity check. And can easily
be removed by reverting this change.

* Disable deprecation check

Between scala 2.12 and 2.13 the map methods were deprecated. Since we
compile for both versions, and the new method does not exist in 2.13, we
disable the check

* Fix test reference of Counter

* WIP: run tests in GHA

* Typo in metric name

* When a metric doesnt exist, NoSuchElement will be thrown instead of NullPointer

* Make attempt tracking more obivous it is happening

* Remove deregister logic

* Use `Long` when getting context fields

* Pass in the task context when getting the counter

* Use safer characters for metrics name

* Wrap parser in a safer construct