Allow control of how many outstanding raw requests there are
Before this fix, all output raw rows for a partition were converted
into queued HTTP requests. This meant that not only did we use memory
for the input rows as part of the spark input dataframes, but also for
the serialized HTTP body versions that were waiting to be sent.. If
then there is a large amount of rows in a partition, or if we are slow
to write to raw (such as for 429s etc), then we end up consuming a lot
of memory intermittently, leading to an OOM (confirmed by analyzing
heap dumps).

After this fix, a new setting, maxOutstandingRawInsertRequests, is
exposed through dataframe options, which controls how many requests we
prepare awaiting execution. This allows us to test this on staging
only, and then roll out for the rest of the clusters. The associated
test is a bit rubbish, but it just verifies that both work (as it
cannot verify that we actually throttled the requests as it is hidden
within the dataframe interface -- nor how many requests are done).

Note that this is potentially made slightly worse that, as far as I
can see, we don't't have a max concurrent writes to Raw at the same
time (as the throttled backend isn't used on write), which means we
will probably hit 429s more frequently (and backoffs will make it use
more time before it gets data out of the heap).

The implementation is not optimal, as noted in the code comment, but I
had issues with cats.effect not wanting to cooperate.